OCT-111: moved GetCurrenciesQuery and IsCurrencyActivatedQuery
