+++ components/catalogs/back/src/Infrastructure/Controller/Internal/GetCurrenciesAction.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Akeneo\Catalogs\Infrastructure\Controller\Internal;
 
-use Akeneo\Catalogs\Application\Persistence\GetCurrenciesQueryInterface;
+use Akeneo\Catalogs\Application\Persistence\Currency\GetCurrenciesQueryInterface;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\RedirectResponse;
 use Symfony\Component\HttpFoundation\Request;
