+++ components/catalogs/back/tests/Integration/Infrastructure/Persistence/GetCurrenciesQueryTest.php
@@ -4,8 +4,8 @@ declare(strict_types=1);
 
 namespace Akeneo\Catalogs\Test\Integration\Infrastructure\Persistence;
 
-use Akeneo\Catalogs\Application\Persistence\GetCurrenciesQueryInterface;
-use Akeneo\Catalogs\Infrastructure\Persistence\GetCurrenciesQuery;
+use Akeneo\Catalogs\Application\Persistence\Currency\GetCurrenciesQueryInterface;
+use Akeneo\Catalogs\Infrastructure\Persistence\Currency\GetCurrenciesQuery;
 use Akeneo\Catalogs\Test\Integration\IntegrationTestCase;
 
 /**
@@ -13,7 +13,7 @@ use Akeneo\Catalogs\Test\Integration\IntegrationTestCase;
  * @copyright 2022 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  *
- * @covers \Akeneo\Catalogs\Infrastructure\Persistence\GetCurrenciesQuery
+ * @covers \Akeneo\Catalogs\Infrastructure\Persistence\Currency\GetCurrenciesQuery
  */
 class GetCurrenciesQueryTest extends IntegrationTestCase
 {
